[ALPS06702793] ffa: dmabuf: fix the 64bit handle
ffa handle is 64bit, so that fix the return handle

MTK-Commit-Id: ec4cee077f82857280bb36a23b748414694f7c3e

CR-Id: ALPS06702793
Feature: devmpu

+++ drivers/dma-buf/heaps/mtk_heap.h
@@ -15,7 +15,7 @@
 extern atomic64_t dma_heap_normal_total;
 
 /* return 0 means error */
-u32 dmabuf_to_secure_handle(const struct dma_buf *dmabuf);
+u64 dmabuf_to_secure_handle(const struct dma_buf *dmabuf);
 
 int is_system_heap_dmabuf(const struct dma_buf *dmabuf);
 int is_mtk_mm_heap_dmabuf(const struct dma_buf *dmabuf);
